Update tests/benchmarks for simple pipelined operations.

+++ Tests/smallthunk.py
@@ -1,8 +1,8 @@
 
 import thunklib
 import numpy
-a = numpy.arange(5, dtype=numpy.int32)
+a = numpy.arange(5, dtype=numpy.double)
 b = thunklib.thunk(a)
-c = b * b
+c = b * b 
 print(type(c))
 print(c)
